backend/requirements.txt
# backend/requirements.txt
flask
flask_cors
gunicorn  # For production deployment
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
APScheduler